+++ setupTests.js
@@ -1,5 +1,5 @@
 /* eslint import/no-extraneous-dependencies:0 */
-import { configure } from 'enzyme';
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
+import { configure } from 'enzyme'
+import Adapter from '@wojtekmaj/enzyme-adapter-react-17'
 
-configure({ adapter: new Adapter() });
+configure({ adapter: new Adapter() })