+++ src/components/EntrySearch.js
@@ -8,6 +8,7 @@ import { Dropdown } from 'primereact/dropdown'
 import { InputText } from 'primereact/inputtext'
 import { Button } from 'primereact/button'
 import { useState } from 'react'
+import { ListBox } from 'primereact/listbox'
 import actions from '../store/actions'
 import { OPERATION_TYPE_DELETE, OPERATION_TYPE_OBJECT_ADD, OPERATION_TYPE_UPDATE } from '../constants/store'
 import SearchBar from './SearchBar'
@@ -20,7 +21,8 @@ const EntrySearch = ({
   upperOntologySearch,
   updateStoreValue,
   annotationProperties,
-  advancedSearchFilters
+  advancedSearchFilters,
+  isFirstQuery
 }) => {
   const { t } = useTranslation()
 
@@ -58,11 +60,15 @@ const EntrySearch = ({
       </h1>
 
       <div className="sidebar-main-body entry-search">
-        <div
-          className="entry-search-row"
-        >
-          <SearchBar />
-        </div>
+        {
+          !isFirstQuery && (
+            <div
+              className="entry-search-row"
+            >
+              <SearchBar />
+            </div>
+          )
+        }
 
         <div
           className="sidebar-main-body-title"
@@ -135,12 +141,21 @@ const EntrySearch = ({
 
                 const isWithEnumeration = ENUMERATION_PROPERTIES.includes(property)
 
+                const enumerationSuggestions = suggestions && suggestions.length > 0
+                  ? suggestions.filter((suggestion) => suggestion.value.toLowerCase().includes(value.toLowerCase()))
+                  : []
+
+                const diplayedEnumerationSuggestions = enumerationSuggestions.length === 1
+                  && enumerationSuggestions[0].value === value
+                  ? []
+                  : enumerationSuggestions
+
                 return (
                   <div
                     key={`advanced-search-${searchFilterKey}`}
                     className="entry-search-block p-pt-3 p-pb-3 p-d-flex p-ai-center"
                   >
-                    <div className="p-d-flex p-flex-column full-width">
+                    <div className="p-d-flex p-flex-column">
                       <div className="entry-search-block-row m-b-5">
                         <Dropdown
                           id={`advanced-search-property-${searchFilterKey}`}
@@ -170,27 +185,23 @@ const EntrySearch = ({
                       </div>
 
                       <div className="entry-search-block-row">
+                        <InputText
+                          className="property-text-input value-input"
+                          id={`advanced-search-value-${searchFilterKey}`}
+                          value={value}
+                          placeholder={t('insertTextOrSelect')}
+                          onChange={(e) => updateStoreValue(['advancedSearchFilters', searchFilterKey], OPERATION_TYPE_OBJECT_ADD, { value: e.target.value })}
+                        />
                         {
-                          isWithEnumeration ? (
-                            <Dropdown
-                              id={`advanced-search-value-${searchFilterKey}`}
+                          isWithEnumeration
+                          && diplayedEnumerationSuggestions.length > 0 && (
+                            <ListBox
                               value={value}
-                              options={suggestions}
-                              filter
+                              options={diplayedEnumerationSuggestions}
                               onChange={(e) => updateStoreValue(['advancedSearchFilters', searchFilterKey], OPERATION_TYPE_OBJECT_ADD, { value: e.value })}
-                              placeholder={t('selectProperty')}
-                            />
-                          ) : (
-                            <InputText
-                              className="property-text-input value-input"
-                              id={`advanced-search-value-${searchFilterKey}`}
-                              value={value}
-                              placeholder={t('insertText')}
-                              onChange={(e) => updateStoreValue(['advancedSearchFilters', searchFilterKey], OPERATION_TYPE_OBJECT_ADD, { value: e.target.value })}
                             />
                           )
                         }
-
                       </div>
                     </div>
 
@@ -258,6 +269,7 @@ EntrySearch.propTypes = {
   updateStoreValue: PropTypes.func.isRequired,
   advancedSearchFilters: PropTypes.shape().isRequired,
   annotationProperties: PropTypes.arrayOf(PropTypes.shape).isRequired,
+  isFirstQuery: PropTypes.bool.isRequired,
 }
 
 const mapToProps = ({
@@ -265,11 +277,13 @@ const mapToProps = ({
   upperOntologySearch,
   advancedSearchFilters,
   annotationProperties,
+  isFirstQuery
 }) => ({
   dataTypeSearch,
   upperOntologySearch,
   advancedSearchFilters,
   annotationProperties,
+  isFirstQuery
 })
 
 export default connect(